Simplify example builds
