+++ examples/react-babel-karma-gulp-fork-ts-checker/webpack.config.prod.js
@@ -1,0 +1,27 @@
+'use strict';
+
+var webpack = require('webpack');
+var webpackFailPlugin = require('webpack-fail-plugin');
+var webpackConfig = require('./webpack.config.base.js');
+
+module.exports = function() {
+  var myProdConfig = webpackConfig;
+  myProdConfig.output.filename = '[name].[hash].js';
+
+  myProdConfig.plugins = myProdConfig.plugins.concat(
+    new webpack.DefinePlugin({
+      'process.env': {
+        'NODE_ENV': JSON.stringify('production')
+      }
+    }),
+    new webpack.optimize.CommonsChunkPlugin({ name: 'vendor', filename: 'vendor.[hash].js' }),
+    new webpack.optimize.UglifyJsPlugin({
+      compress: {
+        warnings: true
+      }
+    }),
+    webpackFailPlugin
+  );
+
+  return myProdConfig;
+};
