+++ examples/react-babel-karma-gulp-happypack/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "karma start --reporters mocha,junit --single-run --browsers PhantomJS",
-    "serve": "gulp watch-and-serve",
+    "serve": "gulp serve",
     "watch": "gulp watch",
     "build": "gulp build"
   },
@@ -43,7 +43,7 @@
     "del": "^2.0.2",
     "express": "^4.13.3",
     "fork-ts-checker-notifier-webpack-plugin": "^0.1.3",
-    "fork-ts-checker-webpack-plugin": "^0.2.0",
+    "fork-ts-checker-webpack-plugin": "^0.2.2",
     "glob": "^7.0.0",
     "gulp": "^3.9.0",
     "gulp-autoprefixer": "^3.1.0",
